requirements.txt
avro-python3==1.8.2
backports.weakref==1.0.post1
certifi==2018.8.24
chardet==3.0.4
confluent-kafka==0.11.5
dictproxyhack==1.1
enum34==1.1.6
funcsigs==1.0.2
futures
gast==0.2.0
grpcio==1.16.0
h5py==2.8.0
idna==2.7
kafka-python==1.4.3
kazoo==2.5.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
mock==2.0.0
numpy==1.15.4
picamera==1.13
pbr==5.1.1
protobuf==3.6.1
pykafka==2.8.0
pyserial==3.4
requests==2.19.1
six==1.11.0
systemd==0.16.1
tabulate==0.8.2
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
urllib3==1.23
Werkzeug==0.14.1
